move graph to graph.special and rearrange tests
