+++ tests/test_algorithms/test_evaluate.py
@@ -11,6 +11,7 @@ from causalprog.graph import (
     DistributionNode,
     Graph,
 )
+from causalprog.graph.special import example_model
 
 
 @pytest.mark.parametrize(
@@ -77,3 +78,28 @@ def test_evaluate_algorithm_four_node():
 
     assert np.isclose(evaluate(graph, "c", a=2.0, b=1.5), 1.5)
     assert np.isclose(evaluate(graph, "x", a=2.0, b=1.5), 4.5)
+
+
+def test_example_model():
+    graph = example_model(
+        compute_u_x=lambda C: C,
+        compute_u_y=lambda C: C + 1,
+        compute_phi_x=lambda L: L[0],
+        compute_x=lambda Z, PhiX, UX: Z[0] + UX - PhiX,
+        compute_y=lambda X, UY: X * UY,
+    )
+
+    data = {
+        "L": np.array([5.5]),
+        "Z": np.array([2.0]),
+        "C": 4.0,
+    }
+
+    assert np.allclose(evaluate(graph, "L", **data), np.array([5.5]))
+    assert np.allclose(evaluate(graph, "Z", **data), np.array([2.0]))
+    assert np.allclose(evaluate(graph, "C", **data), 4.0)
+    assert np.allclose(evaluate(graph, "UX", **data), 4.0)
+    assert np.allclose(evaluate(graph, "UY", **data), 5.0)
+    assert np.allclose(evaluate(graph, "PhiX", **data), 5.5)
+    assert np.allclose(evaluate(graph, "X", **data), 0.5)
+    assert np.isclose(evaluate(graph, "Y", **data), 2.5)
